requirements.txt
Flask==2.0.1
google-cloud-aiplatform==1.13.0
transformers==4.10.0
torch==1.9.0
